+++ releaf-core/app/controllers/releaf/base_controller.rb
@@ -33,12 +33,12 @@ module Releaf
 
     def search(text)
       return if text.blank?
-      return if @searchable_fields.blank?
-      @collection = searcher_class.prepare(relation: @collection, fields: @searchable_fields, text: text)
+      return if searchable_fields.blank?
+      @collection = searcher_class.prepare(relation: @collection, fields: searchable_fields, text: text)
     end
 
     def searcher_class
-      Releaf::Search
+      Releaf::Core::Search
     end
 
     def index
@@ -413,8 +413,12 @@ module Releaf
         index:             true,
         toolbox:           true
       }
-      @panel_layout      = true
-      @resources_per_page    = 40
+      @panel_layout = true
+      @resources_per_page = 40
+    end
+
+    def searchable_fields
+      @searchable_fields ||= Releaf::Core::DefaultSearchableFields.new(resource_class).find
     end
 
     def resource_params