+++ releaf-core/lib/generators/dummy/templates/controllers/admin/authors_controller.rb
@@ -1,9 +1,12 @@
 class Admin::AuthorsController < Releaf::BaseController
+  def searchable_fields
+    [:name, :surname, books: [:title, chapters: [:title]]]
+  end
+
   protected
 
   def setup
     super
-    @searchable_fields = [:name, :surname, books: [:title, chapters: [:title]]]
     @resources_per_page = params.has_key?(:show_all) ? nil : 20
   end
 end