+++ releaf-i18n_database/app/controllers/releaf/i18n_database/translations_controller.rb
@@ -149,7 +149,6 @@ module Releaf::I18nDatabase
       @features = {
         index: true
       }
-      @searchable_fields = true
     end
 
     def action_features